small fix
fixed a odd behavior with pause menu,  if quit from main scene and
started a new. the pause value was not reset on the 2nd load.
